refactor(core): extract sleep+makeAbortError helpers; log clone cwd in observer
